refactor: Stack model
